% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{extract_rho}
\alias{extract_rho}
\title{Extract fitted rho values from model test results}
\usage{
extract_rho(res)
}
\arguments{
\item{res}{A list containing test results for multiple models, where \code{test_res} is a list of individual model test results.}
}
\value{
A data frame containing the fitted rho values for each gene pair.
}
\description{
This function extracts the fitted rho values for each gene pair from a list of model test results.
}
\examples{
# Assuming `res` is a list containing model test results:
fitted_rho_values <- extract_rho(res)
}
