% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_modified.R
\name{fix.family.link.family}
\alias{fix.family.link.family}
\title{Redefine fix.family.link.family to add derivatives to a family object}
\usage{
fix.family.link.family(fam)
}
\arguments{
\item{fam}{A family object for which to add derivatives.}
}
\value{
A modified family object with additional components for derivatives.
}
\description{
This function adds the second, third, and fourth derivatives of the link function
with respect to mu to the family object, which is used for Newton-like optimization.
}
