% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data for spCorr Package}
\format{
A list with the following components:
\describe{
\item{count_mat}{A matrix of raw counts}
\item{count_mat}{A matrix of normalized counts}
\item{gene_list}{A list of genes}
\item{gene_pair_list}{A list of gene pairs}
\item{cov_mat}{A matrix of covariates}
}
}
\usage{
data(test_data)
}
\description{
A dataset containing example data to be used in examples and testing of the functions in the spCorr package.
}
\keyword{datasets}
