% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_models.R
\name{test_models}
\alias{test_models}
\title{Testing function for a list of models}
\usage{
test_models(model_list, ncores = ncores, local_testing = local_testing)
}
\arguments{
\item{model_list}{A list of fitted Generalized Additive Model (GAM) objects to be tested.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing.}

\item{local_testing}{Logical; if \code{TRUE}, performs local significance testing for each model. Default is \code{FALSE}.}
}
\value{
A list containing test results for each model in the input \code{model_list}.
Each element of the list includes:
\describe{
\item{global_test}{A list containing global p-values (\code{global_p}) and effective degrees of freedom (\code{edf}).}
\item{local_test}{A data frame containing local p-values and other local test metrics, if \code{local_testing = TRUE}.}
}
}
\description{
This function performs significance testing on a list of fitted models using parallel processing.
It internally calls \code{test_model()} for each model in the provided list.
}
\examples{
data(test_data)

# Fit standardized marginals for gene expressions
marginals <- fit_marginals(
  gene_list = test_data$gene_list,
  count_mat = test_data$count_mat,
  cov_mat = test_data$cov_mat,
  formula1 = "layer_annotations",
  family1 = "nb",
  to = "gaussian",
  DT = TRUE,
  ncores = 2
)

# Fit product distributions for gene pairs
model_list <- fit_products(
  gene_pair_list = test_data$gene_pair_list,
  marginals = marginals,
  cov_mat = test_data$cov_mat,
  formula2 = "s(x1, x2, bs='tp', k=50)",
  family2 = quasiproductr(),
  control = list(),
  ncores = 2,
  preconstruct_smoother = TRUE
)

# Perform significance testing on the fitted models
test_res <- test_models(
  model_list = model_list,
  ncores = 2,
  local_testing = FALSE
)

}
