% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiproductr.R
\name{quasiproductr}
\alias{quasiproductr}
\title{Define a custom family called 'quasiproductr' to be used with the mgcv package.}
\usage{
quasiproductr(link = "artanh")
}
\arguments{
\item{link}{The link function to be used. Default is \code{"artanh"}.}
}
\value{
A family object compatible with mgcv, including variance, link functions, and residual deviance calculation.
}
\description{
The \code{quasiproductr} family is a custom family added to the \code{mgcv} package to model
relationships where the link function is the hyperbolic arctangent (\code{artanh}).
This family includes functions to define the link, variance, residuals, and
initialization logic required for fitting with Generalized Additive Models (GAM).
}
\examples{
# Example of using quasiproductr with mgcv's gam function
library(mgcv)
fit <- gam(y ~ s(x), family = quasiproductr(), data = my_data)
}
